+++ src/src/Commands/Time.cpp
@@ -83,7 +83,7 @@ String Command_SetDateTime(struct EventStruct *event, const char* Line)
 	}else  {
 		serialPrintln();
 		String result = F("Datetime:");
-		//result += getDateTimeString('-',':',' ');
+		result += getDateTimeString('-',':',' ');
 		return return_result(event, result);
 	}
 	return return_command_success();
