+++ src/src/Commands/Time.h
@@ -3,8 +3,6 @@
 
 class String;
 
-#include <time.h>
-
 String Command_NTPHost(struct EventStruct *event, const char* Line);
 String Command_useNTP(struct EventStruct *event, const char* Line);
 String Command_TimeZone(struct EventStruct *event, const char* Line);
